feat: pure-js sql client agnostic base lib
