+++ src/lib/PostgresMetaFunctions.ts
@@ -1,7 +1,7 @@
 import { ident, literal } from 'pg-format'
 import { DEFAULT_SYSTEM_SCHEMAS } from './constants.js'
 import { filterByList } from './helpers.js'
-import { functionsSql } from './sql/index.js'
+import functionsSql from './sql/functions.sql'
 import { PostgresMetaResult, PostgresFunction, PostgresFunctionCreate } from './types.js'
 
 export default class PostgresMetaFunctions {
