+++ src/lib/base.ts
@@ -1,0 +1,29 @@
+export { default as PostgresMetaBase } from './PostgresMetaBase.js'
+export { wrapError, wrapResult } from './helpers.js'
+export {
+  EndFn,
+  PostgresColumn,
+  PostgresConfig,
+  PostgresExtension,
+  PostgresFunction,
+  PostgresFunctionCreate,
+  PostgresIndex,
+  PostgresMaterializedView,
+  PostgresMetaErr,
+  PostgresMetaOk,
+  PostgresMetaResult,
+  PostgresPolicy,
+  PostgresPrimaryKey,
+  PostgresPublication,
+  PostgresRelationship,
+  PostgresRole,
+  PostgresSchema,
+  PostgresSchemaCreate,
+  PostgresSchemaUpdate,
+  PostgresTable,
+  PostgresTrigger,
+  PostgresType,
+  PostgresVersion,
+  PostgresView,
+  QueryFn,
+} from './types.js'
