+++ src/lib/db.ts
@@ -1,7 +1,7 @@
 import pg, { PoolConfig } from 'pg'
-import { DatabaseError } from 'pg-protocol'
 import { parse as parseArray } from 'postgres-array'
-import { PostgresMetaResult } from './types.js'
+import { wrapError, wrapResult } from './helpers.js'
+import { EndFn, QueryFn } from './types.js'
 
 pg.types.setTypeParser(pg.types.builtins.INT8, (x) => {
   const asNumber = Number(x)
@@ -21,8 +21,8 @@ pg.types.setTypeParser(600, (x) => x) // point
 pg.types.setTypeParser(1017, (x) => x) // _point
 
 export const init: (config: PoolConfig) => {
-  query: (sql: string) => Promise<PostgresMetaResult<any>>
-  end: () => Promise<void>
+  query: QueryFn
+  end: EndFn
 } = (config) => {
   // node-postgres ignores config.ssl if any of sslmode, sslca, sslkey, sslcert,
   // sslrootcert are in the connection string. Here we allow setting sslmode in
@@ -76,77 +76,9 @@ export const init: (config: PoolConfig) => {
         if (Array.isArray(res)) {
           res = res.reverse().find((x) => x.rows.length !== 0) ?? { rows: [] }
         }
-        return { data: res.rows, error: null }
-      } catch (error: any) {
-        if (error instanceof DatabaseError) {
-          // Roughly based on:
-          // - https://github.com/postgres/postgres/blob/fc4089f3c65a5f1b413a3299ba02b66a8e5e37d0/src/interfaces/libpq/fe-protocol3.c#L1018
-          // - https://github.com/brianc/node-postgres/blob/b1a8947738ce0af004cb926f79829bb2abc64aa6/packages/pg/lib/native/query.js#L33
-          let formattedError = ''
-          {
-            if (error.severity) {
-              formattedError += `${error.severity}:  `
-            }
-            if (error.code) {
-              formattedError += `${error.code}: `
-            }
-            if (error.message) {
-              formattedError += error.message
-            }
-            formattedError += '\n'
-            if (error.position) {
-              // error.position is 1-based
-              const position = Number(error.position) - 1
-
-              let line = ''
-              let lineNumber = 0
-              let lineOffset = 0
-
-              const lines = sql.split('\n')
-              let currentOffset = 0
-              for (let i = 0; i < lines.length; i++) {
-                if (currentOffset + lines[i].length > position) {
-                  line = lines[i]
-                  lineNumber = i + 1 // 1-based
-                  lineOffset = position - currentOffset
-                  break
-                }
-                currentOffset += lines[i].length + 1 // 1 extra offset for newline
-              }
-              formattedError += `LINE ${lineNumber}: ${line}
-${' '.repeat(5 + lineNumber.toString().length + 2 + lineOffset)}^
-`
-            }
-            if (error.detail) {
-              formattedError += `DETAIL:  ${error.detail}
-`
-            }
-            if (error.hint) {
-              formattedError += `HINT:  ${error.hint}
-`
-            }
-            if (error.internalQuery) {
-              formattedError += `QUERY:  ${error.internalQuery}
-`
-            }
-            if (error.where) {
-              formattedError += `CONTEXT:  ${error.where}
-`
-            }
-          }
-
-          return {
-            data: null,
-            error: {
-              ...error,
-              // error.message is non-enumerable
-              message: error.message,
-              formattedError,
-            },
-          }
-        }
-
-        return { data: null, error: { message: error.message } }
+        return wrapResult(res.rows)
+      } catch (error) {
+        return wrapError(error, sql)
       }
     },
