+++ src/lib/index.ts
@@ -1,26 +1,3 @@
+export * from './base.js'
+
 export { default as PostgresMeta } from './PostgresMeta.js'
-export {
-  PostgresMetaOk,
-  PostgresMetaErr,
-  PostgresMetaResult,
-  PostgresColumn,
-  PostgresConfig,
-  PostgresExtension,
-  PostgresFunction,
-  PostgresFunctionCreate,
-  PostgresIndex,
-  PostgresMaterializedView,
-  PostgresPolicy,
-  PostgresPrimaryKey,
-  PostgresPublication,
-  PostgresRelationship,
-  PostgresRole,
-  PostgresSchema,
-  PostgresSchemaCreate,
-  PostgresSchemaUpdate,
-  PostgresTable,
-  PostgresTrigger,
-  PostgresType,
-  PostgresVersion,
-  PostgresView,
-} from './types.js'
