+++ src/types/sql.d.ts
@@ -1,0 +1,4 @@
+declare module '*.sql' {
+  const sql: string;
+  export default sql;
+}
